requirements.txt
aiohttp==3.7.3
async-timeout==3.0.1
attrs==20.3.0
bidict==0.21.2
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
dnspython==1.16.0
eventlet==0.30.0
greenlet==0.4.17
gunicorn==20.0.4
h11==0.11.0
httptools==0.1.1
idna==2.10
multidict==5.1.0
python-dotenv==0.15.0
python-engineio==4.0.0
python-socketio==5.0.4
PyYAML==5.3.1
requests==2.25.1
six==1.15.0
typing-extensions==3.7.4.3
urllib3==1.26.2
uvicorn==0.13.2
uvloop==0.14.0
watchgod==0.6
websocket-client==0.57.0
websockets==8.1
yarl==1.6.3
